+++ packages/snap-preact-demo/tests/cypress/e2e/templates/recommendation/recommendation.cy.js
@@ -181,18 +181,19 @@ describe('Recommendations', () => {
 				});
 
 				it('can click on a result and go to that page', function () {
-					cy.document().then((doc) => {
-						cy.snapController(integration?.selectors?.recommendation.controller).then(({ store }) => {
-							cy.get(integration?.selectors?.recommendation.activeSlide).should('exist');
-							let url = doc.querySelector(`${integration?.selectors?.recommendation.activeSlide} ${integration?.selectors?.recommendation.result} a`)
-								.attributes?.href?.value;
-							cy.get(integration?.selectors?.recommendation.activeSlide)
-								.click({ multiple: true })
+					cy.get(integration?.selectors?.recommendation.activeSlide).should('exist');
+					cy.get(`${integration?.selectors?.recommendation.activeSlide} ${integration?.selectors?.recommendation.result} a`)
+						.first()
+						.invoke('attr', 'href')
+						.then((url) => {
+							expect(url).to.be.a('string').and.not.be.empty;
+							cy.get(`${integration?.selectors?.recommendation.activeSlide} ${integration?.selectors?.recommendation.result} a`)
+								.first()
+								.click({ force: true })
 								.then(() => {
 									cy.location('href').should('include', url);
 								});
 						});
-					});
 				});
 
 				describe('Tests Custom Result Component', () => {