+++ packages/snap-preact/src/Templates/Stores/library/languages/es.ts
@@ -113,23 +113,20 @@ export const es: LangComponents = {
 			value: (data) => `Subtotal para ${data.cartStore.count} artículos`,
 		},
 	},
-	mobileSidebar: {
-		openButtonText: {
-			value: 'Filtros',
+	result: {
+		addToCartButtonText: {
+			value: 'Añadir todo al carrito',
 		},
-		clearButtonText: {
-			value: 'Limpiar todo',
+		addToCartButtonSuccessText: {
+			value: '¡Agregado!',
 		},
-		applyButtonText: {
-			value: 'Aplicar',
+	},
+	overlayResult: {
+		addToCartButtonText: {
+			value: 'Añadir todo al carrito',
 		},
-		titleText: {
-			value: 'Opciones de filtro',
-		},
-		closeButtonText: {
-			attributes: {
-				'aria-label': 'cerrar botón de filtros',
-			},
+		addToCartButtonSuccessText: {
+			value: '¡Agregado!',
 		},
 	},
 	slideshow: {