+++ packages/snap-preact-demo/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0](https://github.com/AthosCommerce/snap/compare/v1.7.0...v1.8.0) (2026-07-13)
+
+### Features
+
+- **preact/templates:** adding globalResultComponent configuration for convenience ([bdee4ae](https://github.com/AthosCommerce/snap/commit/bdee4ae856c2477a5d6b2b6331882eec56946e5f))
+
 # [1.7.0](https://github.com/AthosCommerce/snap/compare/v1.6.0...v1.7.0) (2026-06-11)
 
 **Note:** Version bump only for package @athoscommerce/snap-preact-demo